+++ app/(components)/layout/CookieConsent/components/CookieConsentPopover.tsx
@@ -1,0 +1,34 @@
+'use client';
+
+import Link from 'next/link';
+
+import { IntlText } from '@/components/intl';
+import { Button, Typography, typographyVariants } from '@/components/ui';
+
+import { useCookieConsent } from '../hooks';
+
+export const CookieConsentPopover = () => {
+  const { accept } = useCookieConsent();
+
+  return (
+    <div className='fixed bottom-0 left-0 z-50 m-5'>
+      <div className='flex max-w-120 flex-col gap-6 rounded-16 border border-border-hard bg-white p-6 dark:bg-secondary'>
+        <Typography as='p' variant='body-sm'>
+          <IntlText
+            values={{
+              link: (chunks) => (
+                <Link className={typographyVariants({ variant: 'link' })} href='#'>
+                  {chunks}
+                </Link>
+              )
+            }}
+            path='cookieConsent.description'
+          />
+        </Typography>
+        <Button size='lg' onClick={accept}>
+          <IntlText path='button.accept' />
+        </Button>
+      </div>
+    </div>
+  );
+};
