+++ app/(components)/layout/CookieConsent/hooks/useCookieConsent.ts
@@ -1,0 +1,17 @@
+import { useCookie } from '@siberiacancode/reactuse';
+
+import { COOKIES } from '@/app/(constants)';
+
+const PATH = '/';
+
+export const useCookieConsent = () => {
+  const cookieConsent = useCookie(COOKIES.COOKIE_CONSENT, {
+    path: PATH,
+    initialValue: false
+  });
+
+  return {
+    value: cookieConsent.value,
+    accept: () => cookieConsent.set(true, { path: PATH })
+  };
+};
