+++ app/(components)/layout/CookieConsent/scripts/TrackingScripts/TrackingScripts.tsx
@@ -1,0 +1,9 @@
+import { GoogleTagManagerScript } from './GoogleTagManagerScript';
+import { YandexMetrikaScript } from './YandexMetricaScript';
+
+export const TrackingScripts = () => (
+  <>
+    <GoogleTagManagerScript />
+    <YandexMetrikaScript />
+  </>
+);
