+++ app/(components)/layout/CookieConsent/scripts/index.ts
@@ -1,0 +1,1 @@
+export * from './TrackingScripts/TrackingScripts';
