+++ app/(components)/scripts/index.ts
@@ -1,3 +1,1 @@
-export * from './GoogleTagManagerScript/GoogleTagManagerScript';
 export * from './ThemeScript/ThemeScript';
-export * from './YandexMetricaScript/YandexMetricaScript';
